Added responsiveness
