+++ mitushconnect/src/App.js
@@ -1,9 +1,11 @@
- import LandingPage from "./LandingPage"
+import Navbar from "./Navbar";
+import LandingPage from "./LandingPage";
 import './App.css';
 
 function App() {
   return (
     <div >
+       <Navbar/>
        <LandingPage/>
     </div>
   );
